First skeleton for the v2 added.
Changed to promises.

Dropped (for now) the "srcHost" param because it complicates a lot the code. We need a simpler solution.

Simpler tests. Engine changed to "tap", also simple, but a bit more needed features (i.e.: code completion).

One module refactored ("geolocation") and its tests added.

A lot of things still broken and not finished.